requirements.txt
fastapi==0.128.3
uvicorn[standard]==0.40.0
python-dotenv==1.2.1
fastmcp==2.14.5


# LLM and AI
groq==0.37.1
langchain-groq==1.1.2
langchain-core==1.2.9
langchain-ollama==1.0.1
langchain-mcp-adapters==0.2.1
langgraph==1.0.8
langgraph-checkpoint-sqlite==3.0.3

chromadb==1.4.1
llama-index==0.14.13
llama-index-embeddings-huggingface
llama-index-llms-ollama  # Commented out: using Groq Cloud instead
llama-index-llms-groq
llama-index-readers-web==0.5.6
llama-index-readers-file==0.5.6
llama-index-vector-stores-chroma

# Evaluation
deepeval==3.8.4